refactor: minor ui changes to logout modal
