+++ src/components/Modals/LogoutModal/index.jsx
@@ -1,0 +1,2 @@
+import LogoutModal from "./LogoutModal";
+export default LogoutModal;
